+++ analysis_driver/pipelines/demultiplexing.py
@@ -1,7 +1,11 @@
 import shutil
 from os import mkdir
-from os.path import join, exists, isdir
+from os.path import join, exists, isdir, basename, dirname
+import os
 from egcg_core.config import cfg
+from egcg_core import executor, util, clarity
+from egcg_core.constants import ELEMENT_PROJECT_ID, ELEMENT_LANE, ELEMENT_SAMPLE_INTERNAL_ID
+from egcg_core.util import find_file
 from egcg_core import executor, util, rest_communication
 from analysis_driver import segmentation
 from analysis_driver.util import bash_commands, find_all_fastq_pairs_for_lane, get_trim_values_for_bad_cycles
@@ -212,6 +216,162 @@ class RunReview(DemultiplexingStage):
         return 0
 
 
+class PostDemultiplexingStage(DemultiplexingStage):
+
+    _fastq_files = {}
+
+    @property
+    def alignment_dir(self):
+        return join(self.job_dir, 'alignment')
+
+    def fastq_pair(self, run_element):
+        re_id = (
+            run_element.get(ELEMENT_PROJECT_ID),
+            run_element.get(ELEMENT_SAMPLE_INTERNAL_ID),
+            run_element.get(ELEMENT_LANE)
+        )
+        if re_id not in self._fastq_files:
+            fqs = util.find_fastqs(
+                self.fastq_dir,
+                run_element.get(ELEMENT_PROJECT_ID),
+                run_element.get(ELEMENT_SAMPLE_INTERNAL_ID),
+                run_element.get(ELEMENT_LANE)
+            )
+            fqs.sort()
+            self._fastq_files[re_id] = fqs
+        return self._fastq_files[re_id]
+
+    def fastq_base(self, run_element):
+        fastq_pair = self.fastq_pair(run_element)
+        return fastq_pair[0][:-len('_R1_001.fastq.gz')]
+
+    def bam_path(self, run_element):
+        return join(
+            self.alignment_dir,
+            run_element.get(ELEMENT_PROJECT_ID),
+            run_element.get(ELEMENT_SAMPLE_INTERNAL_ID),
+            basename(self.fastq_base(run_element))
+        ) + '.bam'
+
+
+class BwaAlignMulti(PostDemultiplexingStage):
+    def _run(self):
+        bwa_commands = []
+        self.debug('Searching for fastqs in ' + self.fastq_dir)
+        for run_element in self.dataset.run_elements:
+            if self.fastq_pair(run_element):
+                # make sure the directory where the bam file will go exists
+                os.makedirs(dirname(self.bam_path(run_element)), exist_ok=True)
+                # get the reference genome
+                species = clarity.get_species_from_sample(run_element.get(ELEMENT_SAMPLE_INTERNAL_ID))
+                default_genome_version = cfg.query('species', species, 'default')
+                reference_genome = cfg.query('genomes', default_genome_version, 'fasta')
+                bwa_commands.append(
+                    bash_commands.bwa_mem_biobambam(
+                        self.fastq_pair(run_element),
+                        reference_genome,
+                        self.bam_path(run_element),
+                        {'ID': '1', 'SM': run_element.get(ELEMENT_SAMPLE_INTERNAL_ID), 'PL': 'illumina'},
+                        thread=6
+                    )
+                )
+        return executor.execute(
+            *bwa_commands,
+            job_name='bwa_mem',
+            working_dir=self.job_dir,
+            cpus=4,
+            mem=24,
+            log_commands=False
+        ).join()
+
+
+class SamtoolsStatsMulti(PostDemultiplexingStage):
+    def _run(self):
+        samtools_stats_cmds = []
+        for run_element in self.dataset.run_elements:
+            if self.fastq_pair(run_element):
+                samtools_stats_cmds.append(bash_commands.samtools_stats(
+                    self.bam_path(run_element),
+                    self.fastq_base(run_element) + '_samtools_stats.txt'
+                ))
+        return executor.execute(
+            *samtools_stats_cmds,
+            job_name='samtoolsstats',
+            working_dir=self.job_dir,
+            cpus=1,
+            mem=8,
+            log_commands=False
+        ).join()
+
+
+class SamtoolsDepthMulti(PostDemultiplexingStage):
+
+    def _run(self):
+        samtools_depth_cmds = []
+        for run_element in self.dataset.run_elements:
+            if self.fastq_pair(run_element):
+                samtools_depth_cmds.append(bash_commands.samtools_depth_command(
+                    self.job_dir,
+                    self.bam_path(run_element),
+                    self.fastq_base(run_element) + '_samtools.depth'
+                ))
+        return executor.execute(
+            *samtools_depth_cmds,
+            job_name='samtoolsdepth',
+            working_dir=self.job_dir,
+            cpus=1,
+            mem=6,
+            log_commands=False
+        ).join()
+
+
+class PicardMarkDuplicateMulti(PostDemultiplexingStage):
+    def _run(self):
+        mark_dup_cmds = []
+        for run_element in self.dataset.run_elements:
+            if self.fastq_pair(run_element):
+                out_md_bam = self.bam_path(run_element)[:-len('.bam')] + '_markdup.bam'
+                metrics_file = self.fastq_base(run_element) + '_markdup.metrics'
+                mark_dup_cmds.append(bash_commands.picard_mark_dup_command(
+                    self.bam_path(run_element),
+                    out_md_bam,
+                    metrics_file
+                ))
+        return executor.execute(
+            *mark_dup_cmds,
+            job_name='picardMD',
+            working_dir=self.job_dir,
+            cpus=1,
+            mem=12
+        ).join()
+
+
+class PicardInsertSizeMulti(PostDemultiplexingStage):
+    def _run(self):
+        insert_size_cmds = []
+        for run_element in self.dataset.run_elements:
+            if self.fastq_pair(run_element):
+                metrics_file = self.fastq_base(run_element) + '_insertsize.metrics'
+                histogram_file = self.fastq_base(run_element) + '_insertsize.pdf'
+                insert_size_cmds.append(bash_commands.picard_insert_size_command(
+                    self.bam_path(run_element),
+                    metrics_file,
+                    histogram_file
+                ))
+        return executor.execute(
+            *insert_size_cmds,
+            job_name='picardIS',
+            working_dir=self.job_dir,
+            cpus=1,
+            mem=12
+        ).join()
+
+
+# Need to have a different name of class bcause the sage name is based on it.
+class QCOutput2(QCOutput):
+    pass
+
+
 def build_pipeline(dataset):
 
     def stage(cls, **params):
@@ -226,7 +386,13 @@ def build_pipeline(dataset):
     seqtk = stage(SeqtkFQChk, previous_stages=[fastq_filter])
     md5 = stage(MD5Sum, previous_stages=[fastq_filter])
     qc_output = stage(QCOutput, previous_stages=[welldups, integrity_check, fastqc, seqtk, md5])
-    data_output = stage(DataOutput, previous_stages=[qc_output])
+    align_output  = stage(BwaAlignMulti, previous_stages=[qc_output])
+    stats_output = stage(SamtoolsStatsMulti, previous_stages=[align_output])
+    depth_output = stage(SamtoolsDepthMulti, previous_stages=[align_output])
+    md_output = stage(PicardMarkDuplicateMulti, previous_stages=[align_output])
+    is_output = stage(PicardInsertSizeMulti, previous_stages=[align_output])
+    qc_output2 = stage(QCOutput2, previous_stages=[stats_output, depth_output, md_output, is_output])
+    data_output = stage(DataOutput, previous_stages=[qc_output2])
     _cleanup = stage(Cleanup, previous_stages=[data_output])
     review = stage(RunReview, previous_stages=[_cleanup])
     return review